+++ customize.sh
@@ -12,7 +12,7 @@ SKIPUNZIP=0
 print_modname() {
   ui_print " "
   ui_print "*******************************"
-  ui_print " Supports Android7-15 move cert"
+  ui_print " Supports Android7-16 move cert"
   ui_print "*******************************"
   ui_print " "
 }
@@ -21,8 +21,9 @@ print_modname() {
 
 on_install() {
 
-  # F_TARGETDIR="$MODDIR/system/etc/security/cacerts/"
+  F_TARGETDIR="$MODPATH/system/etc/security/cacerts/"
   D_CERTIFICATE="$MODPATH/certificates"
+  APEX_CONSCRYPT_DIR="$MODPATH/apex/com.android.conscrypt/cacerts"
 
   # mkdir -p "$F_TARGETDIR"
   # create temp cert
@@ -35,6 +36,9 @@ on_install() {
     mkdir -p -m 777 "$D_TMP_CERT"
   fi
   # ui_print "- mkdir $MODPATH/certificates"
+  ui_print "- mkdir $F_TARGETDIR"
+  mkdir -p -m 755 "$F_TARGETDIR"
+  mkdir -p -m 755 "$APEX_CONSCRYPT_DIR"
   mkdir -p -m 755 "$D_CERTIFICATE"
   mkdir -p -m 755 /data/misc/user/0/cacerts-added
 }