+++ sh/refresh.sh
@@ -1,0 +1,4 @@
+#!/system/bin/sh
+MODDIR=${0%/*}
+
+# TODO 待实现